* Fixed validation error and header bar markup. [Bug 4549]
We have several cases in admin where HTML validation is borked if there are no items displayed in a table. The problem is that the php was written without taking into account the possibility of empty tables. I'll fix them as I come across them but wont go looking for them at this stage.

git-svn-id: https://devel.simplemachines.org/code/smf/trunk@10358 a96bbafc-d752-43aa-a6db-eed64a5f2cf9